Delta

+++ include/f_core/os/tenants/c_cpu_monitor_tenant.h
@@ -28,9 +28,8 @@ public:
 
 private:
     CMessagePort<CpuMonitorData>& outputPort;
-
-    uint32_t prevExecutionCycles = 0;
-    uint32_t prevTotalCycles = 0;
+    uint64_t prevExecutionCycles = 0;
+    uint64_t prevTotalCycles = 0;
 
 #if DT_NODE_EXISTS(DT_ALIAS(die_temp))
     const device *dieTempSensor = DEVICE_DT_GET(DT_ALIAS(die_temp));
